feat: custom oranisation order for edit software page
